refactor: 출력 메서드 LottoPrinter 클래스로 분리
